backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary>=2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# LLM - Mistral AI (Free tier available!)
mistralai==1.2.5

# MCP (Model Context Protocol)
mcp==1.3.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0

# Email
sendgrid==6.11.0

# Notifications
slack-sdk==3.27.0

# Utilities
httpx==0.26.0
python-dateutil==2.8.2
pytz==2024.1
